[day07] this works, provides outputs for several runs, but
I do not get the expected result. in this implementation, once we have output
the programm is "paused", the output is passed to the next amplifier as an
input. When the amplifier is reached the next time, the program continues
where it was paused in the previous instance. The program is not reset during
calculation. Therefore, these two values (input vector and position) are
passed as reference.

If the run is marked as initial, the first input instruction sent to the
amplifier is its phase signal, and otherwise it is the input signal. HA!

I GOT IT. The input is called more than once during a single execution, and
I did not account for that with the OptionalInput check in code case == 4.