[day09] add support for program resizing. not working yet:
I end up with a code == 0 at one point. So far, I added the support for
optcode 9, which takes its parameter and increases the value of RelativeBase
by that amount. I am not sure whether 9 can only take paramaters in imediate
mode - that would make sense. If that is the case I need to add a check in
GetParameterMode function for that.