[day05 and day07] make sure that day 5 works with the code from day 07
This is prep for day 9 (which does not work at the moment). The idea is to
move optcode (I know, it is called intcode but I got mixed up) into a separate
file and have it called from code for days 5, 7, and 9. I should also try to
do the same for day 2.