[day16] using those patterns, if we are over half, it means to
just accumulate everything from there until the end. Actually I see
a recursive thing here: the similar pattern may be established for
the second quarter of the input - but then we should not go until
`Pattern.end()` but just for the amount of places corresponding to the
current pass. I have tried that but could not get it to work properly
yet.