[day11] party 2: print white coords
again this was not working as expected because of the way I initialized
the empty list. AGAIN LIST MUTABILITY, when will I learn...

so remember, empty matrix should always be initialized this way!

`[x*[' ']] * y` will result in all rows being modified if one row is
modified because we are pointing to the same object!